perf(core): cythonize Heisenberg algorithm (#5)
Significantly improved serial performance of the Heisenberg model
simulation by converting the main simulation loops into Cython
code. The key changes were as follows,

* Created C++ compatible data containers to mirror the dataclass
  containers for storing the simulation state and lookup tables.
* Use memory views to access numpy arrays within Cython-generated
  C++ code.
* Ported the Metropolis-Hasting algorithm for the Heisenberg model
  simulation to Cython code.
* Ported the order parameter calculations for the Heisenberg model
  to Cython code.
* Implemented wrapper interface to the Mersenne Twister RNG from
  the C++ STL.
* Implemented Python interface for executing the main simulation
  loops now written in Cython code.

The method for gathering statistics and computing order parameters
has been revised to minimize what is computed during a simulation
loop and now supports computations for the heat capacity, the spin
susceptibility, antiferromagnetic ordering, and Binder parameters.
Most of the estimators are no longer computing "on-the-fly" and are
only normalized at the very end of the simulation to reduce
numerical error. Now, the simulation just tracks the total energy,
sums of spins on sublattices, and the total magnetization after
each lattice sweep. A post-processing step at the end of simulation
is used to compute all estimator moments and running averages for
the relevant order parameters.

Package dependencies and project configuration files were updated
to support the use of Cython. Cython was configured to generate C++
code instead of C code.

All Python code in the project will now be formatted using Black.

The Makefile for the project has been streamlined to use a
self-documenting sed script.

The project now uses pre-commit hooks to ensure better code style
and check for typing consistency using mypy.

Closes #5 

BREAKING CHANGE:

The change over to Cython code broke the implementations of the
pure Python versions of the Ising and Heisenberg models, so those
implementations and their corresponding unit tests have been
removed.

This is still a prototype and should not be viewed as a stable
release in any way. The code is still in the alpha phase and I am
planning to significantly refactor the code in order to create a
more stable API.